Fix no songs in queue bug and update to new Html 5 Spec
Before this commit if there were no songs in the queue and the previous
song had ended the player would be stuck trying to pop a song off of the
queue that didn't exist. By checking if the queue is empty before we get
rid of this problem.

This commit also addresses the new Html 5 web audio spec and the removal
of the gain attribute on the source node.